% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create a NOACS project}
\usage{
create(name)
}
\arguments{
\item{name}{Character. The name of your project. No default.}
}
\description{
This functions creates a NOACS project by providing a structure to
your documents and code.
}
