% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_injury_dataset_for_ais_coding.R
\name{create_injury_dataset_for_ais_coding}
\alias{create_injury_dataset_for_ais_coding}
\title{Create an injury dataset for coding using the Abbrevaited Injury Scale (AIS)}
\usage{
create_injury_dataset_for_ais_coding(
  data,
  injury.columns,
  other.columns = NULL,
  only.unique = TRUE,
  remove.coded = TRUE,
  save.dataset = FALSE,
  file = "injury-dataset-for-ais-coding"
)
}
\arguments{
\item{data}{A data.frame. The dataset containing the injury
data. No default.}

\item{injury.columns}{Character. The names of the columns with
injury descriptions. No default.}

\item{other.columns}{Character or NULL. The names of other columns
to be included in the output. Defaults to NULL.}

\item{only.unique}{Logical. If TRUE only unique injuries are kept
in the dataset. Defaults to TRUE.}

\item{remove.coded}{Logical. If TRUE injuries that are already
coded are removed from the dataset. Defaults to TRUE.}

\item{save.dataset}{Logical. If TRUE the created dataset is saved
to disk as a .csv file and returned invisibly. Defaults to
FALSE.}

\item{file}{Character. The path of the file containing the injury
dataset to code. Only used if save.dataset = TRUE and defaults
to "injury-dataset-for-ais-coding".}
}
\description{
Takes a dataset that contains injury data as input and creates a
new dataset for AIS coding
}
