% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_injury_data.R
\name{extract_injury_data}
\alias{extract_injury_data}
\title{Extract injury data}
\usage{
extract_injury_data(
  data,
  injury.columns,
  other.columns = NULL,
  only.unique = TRUE
)
}
\arguments{
\item{data}{A data.frame. The dataset containing the injury
data. No default.}

\item{injury.columns}{Character. The names of the columns with
injury descriptions. No default.}

\item{other.columns}{Character or NULL. The names of other columns
to be included in the output. Defaults to NULL.}

\item{only.unique}{Logical. If TRUE only unique injuries are kept
in the dataset. Defaults to TRUE.}
}
\description{
Extract injury data from a dataset as a single vector
}
