% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hijack_function.R
\name{hijack_function}
\alias{hijack_function}
\title{Hijack function}
\usage{
hijack_function(FUN, ...)
}
\arguments{
\item{FUN}{Function. The function to hijack. No default.}

\item{...}{The arguments to change.}
}
\description{
Creates a new function based on another function but changes one or
more default values. Stolen
from https://trinkerrstuff.wordpress.com/2014/08/19/hijacking-r-functions-changing-default-arguments-3/
}
