% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv_from_github.R
\name{import_csv_from_github}
\alias{import_csv_from_github}
\title{This function imports csv files from GitHub, it will rewrite the GitHub-base URL to the raw format.}
\usage{
import_csv_from_github(ghurl, delim = ",")
}
\arguments{
\item{ghurl}{URL to github csv file, for example https://github.com/titco/titco-I/blob/master/codebook-titco-I-v1.csv}

\item{delim}{The deliminator used in the csv file, either "," or ";"- Defaults to ","}
}
\value{
data A tibble with the data from the csv file.
}
\description{
This function imports csv files from GitHub, it will rewrite the GitHub-base URL to the raw format.
}
