% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_get_project_codebook.R
\name{kobo_get_project_codebook}
\alias{kobo_get_project_codebook}
\title{Functions starting with kobo_ are a set of functions to interact with the datacollection tool kobotoolbox.
Get the XLSForm/Codebook for a kobo project.}
\usage{
kobo_get_project_codebook(url, uid, username, password)
}
\arguments{
\item{url}{URL to kobotoolbox, for example https://kf.kobo.prod.in.noacs.io}

\item{uid}{The uid for with the project, get it from kobo_get_project_list function}

\item{username}{Username for kobotoolbox}

\item{password}{Password for kobotoolbox}
}
\value{
data A list with a tibble for each sheet in the XLSForm.
}
\description{
Functions starting with kobo_ are a set of functions to interact with the datacollection tool kobotoolbox.
Get the XLSForm/Codebook for a kobo project.
}
