% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_get_project_data.R
\name{kobo_get_project_data}
\alias{kobo_get_project_data}
\title{Functions starting with kobo_ are a set of functions to interact with the datacollection tool kobotoolbox.
Get data for a kobo project.}
\usage{
kobo_get_project_data(url, uid, username, password)
}
\arguments{
\item{url}{URL to kobotoolbox, for example https://kf.kobo.prod.in.noacs.io}

\item{uid}{The uid for with the project, get it from kobo_get_project_list function}

\item{username}{Username for kobotoolbox}

\item{password}{Password for kobotoolbox}
}
\value{
data A list with project data, note that some projects may return nestled lists within the result depending on the project configuration.
}
\description{
Functions starting with kobo_ are a set of functions to interact with the datacollection tool kobotoolbox.
Get data for a kobo project.
}
