% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noacsr-package.R
\docType{package}
\name{noacsr-package}
\alias{noacsr}
\alias{noacsr-package}
\title{noacsr: NOACS Research Lab Utility Functions}
\description{
NOACS - the No Accidents research lab conducts medical research on the prediction and prevention of poor patient outcomes. This package is a collection of utility functions for streamlining our research work and facilitate onboarding of new members and collaborators.
}
\author{
\strong{Maintainer}: Martin Gerdin Wärnberg \email{martingerdin@gmail.com} (\href{https://orcid.org/0000-0001-6069-4794}{ORCID})

Authors:
\itemize{
  \item Johanna Berg \email{johanna.berg@ki.se} (\href{https://orcid.org/0000-0001-7553-7337}{ORCID})
}

}
\keyword{internal}
