% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_database_access.R
\name{setup_database_access}
\alias{setup_database_access}
\title{Setup database access}
\usage{
setup_database_access(silent = FALSE)
}
\arguments{
\item{silent}{Logical. If FALSE the function will print messages if
database username and passwords are already found as
environment variables.}
}
\description{
Internal function to setup database access by creating a .env file
for storage of database credentials.
}
