% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_all_functions.R
\name{source_all_functions}
\alias{source_all_functions}
\title{Source all functions}
\usage{
source_all_functions()
}
\description{
Sources all files from the functions directory so that they are available in main.R
}
