% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create a NOACS project}
\usage{
create(name, open.manuscript = TRUE, setup.database.access = FALSE)
}
\arguments{
\item{name}{Character. The name of your project. No default.}

\item{open.manuscript}{Logical. If TRUE the file manuscript.Rmd
created by this function is opened in your current
editor. Defaults to TRUE.}

\item{setup.database.access}{Logical. If TRUE some additional steps
are added to the create project process where the user is asked
to setup database access.}
}
\description{
This functions creates a NOACS project by providing a structure to
your documents and code.
}
