% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_function.R
\name{create_function}
\alias{create_function}
\title{Create Function}
\usage{
create_function(title)
}
\arguments{
\item{title}{Character. What the function does, for example
"Create function" of "Create sample characteristics table". No
default.}
}
\description{
Creates a file in the directory \verb{/functions} with some function
boilerplate to facilitate creating functions and write code that is
easier to understand and maintain.
}
